Deprecate ProjectDependency#getDependencyProject
This method allows the user to fetch the mutable Project instance of a target project.
Gradle projects should not access the mutable state of other projects.
Within Isolated Projects, this behavior is forbidden.

We deprecate this method as a Dependency should only reference the identity of the target proejct and
should not hold onto the mutable reference of the target project. As such, it should also not expose
that mutable reference.